Moved AutoSynchronizationService to UWP specific services.
