+++ src/SilentNotes.Android/Startup.cs
@@ -3,11 +3,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-using Android.App;
 using CommunityToolkit.Mvvm.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection;
 using SilentNotes.Android.Services;
-using SilentNotes.HtmlView;
 using SilentNotes.Services;
 using SilentNotes.Workers;
 
@@ -79,15 +77,7 @@ namespace SilentNotes.Android
                 serviceProvider.GetService<ICryptoRandomService>()));
             services.AddSingleton<IInternetStateService>((serviceProvider) => new InternetStateService(
                 serviceProvider.GetService<IAppContextService>()));
-
-            // todo: remove namespace
-            services.AddSingleton<IAutoSynchronizationService>((serviceProvider) => new SilentNotes.Android.Services.AutoSynchronizationService());
-
-            //services.AddSingleton<IAutoSynchronizationService>((serviceProvider) => new AutoSynchronizationService(
-            //    serviceProvider.GetService<IInternetStateService>(),
-            //    serviceProvider.GetService<ISettingsService>(),
-            //    serviceProvider.GetService<IRepositoryStorageService>(),
-            //    serviceProvider.GetService<INavigationService>()));
+            services.AddSingleton<IAutoSynchronizationService>((serviceProvider) => new AutoSynchronizationService());
             services.AddSingleton<IThemeService>((serviceProvider) => new ThemeService(
                 serviceProvider.GetService<ISettingsService>(),
                 serviceProvider.GetService<IEnvironmentService>()));
