+++ src/SilentNotes.UWP/SilentNotes.UWP.csproj
@@ -113,6 +113,7 @@
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Services\AutoSynchronizationService.cs" />
     <Compile Include="Services\CryptoRandomService.cs" />
     <Compile Include="Services\DataProtectionService.cs" />
     <Compile Include="Services\EnvironmentService.cs" />
